storage/requirements.txt
sqlmodel
fastapi
uvicorn
sqlalchemy_schemadisplay
python-multipart
vention-storage == 0.0.0